organized the old components
