--- apps/nextjs/src/components/landingPage/oldComponents/Faq.tsx
@@ -1,706 +1,0 @@
-export default function Faq() {
-  return (
-    <section className="relative z-20 overflow-hidden bg-white pb-8 pt-20 lg:pb-[50px] lg:pt-[120px] pl-5 lg:pl-0">
-      <div className="container mx-auto">
-        <div className="-mx-4 flex flex-wrap">
-          <div className="w-full px-4">
-            <div className="mx-auto mb-[60px] max-w-[520px] text-center">
-              <span className="mb-2 block text-lg font-semibold text-primary">
-                FAQ
-              </span>
-              <h2 className="mb-3 text-3xl font-bold leading-[1.2] text-black sm:text-4xl md:text-[40px]">
-                Any Questions? Contact Us
-              </h2>
-            </div>
-          </div>
-        </div>
-        <div className="-mx-4 flex flex-wrap">
-          <div className="w-full px-4 lg:w-1/2">
-            <div className="mb-12 flex lg:mb-[70px]">
-              <div className="mr-4 flex h-[50px] w-full max-w-[50px] items-center justify-center rounded-xl bg-primary text-white sm:mr-6 sm:h-[60px] sm:max-w-[60px]">
-                <svg
-                  width="32"
-                  height="32"
-                  viewBox="0 0 34 34"
-                  className="fill-current"
-                >
-                  <path d="M17.0008 0.690674C7.96953 0.690674 0.691406 7.9688 0.691406 17C0.691406 26.0313 7.96953 33.3625 17.0008 33.3625C26.032 33.3625 33.3633 26.0313 33.3633 17C33.3633 7.9688 26.032 0.690674 17.0008 0.690674ZM17.0008 31.5032C9.03203 31.5032 2.55078 24.9688 2.55078 17C2.55078 9.0313 9.03203 2.55005 17.0008 2.55005C24.9695 2.55005 31.5039 9.0313 31.5039 17C31.5039 24.9688 24.9695 31.5032 17.0008 31.5032Z" />
-                  <path d="M17.9039 6.32194C16.3633 6.05631 14.8227 6.48131 13.707 7.43756C12.5383 8.39381 11.8477 9.82819 11.8477 11.3688C11.8477 11.9532 11.9539 12.5376 12.1664 13.0688C12.3258 13.5469 12.857 13.8126 13.3352 13.6532C13.8133 13.4938 14.0789 12.9626 13.9195 12.4844C13.8133 12.1126 13.707 11.7938 13.707 11.3688C13.707 10.4126 14.132 9.50944 14.8758 8.87194C15.6195 8.23444 16.5758 7.96881 17.5852 8.18131C18.9133 8.39381 19.9758 9.50944 20.1883 10.7844C20.4539 12.3251 19.657 13.8126 18.2227 14.3969C16.8945 14.9282 16.0445 16.2563 16.0445 17.7969V21.1969C16.0445 21.7282 16.4695 22.1532 17.0008 22.1532C17.532 22.1532 17.957 21.7282 17.957 21.1969V17.7969C17.957 17.0532 18.382 16.3626 18.9664 16.1501C21.1977 15.2469 22.4727 12.9094 22.0477 10.4657C21.6758 8.39381 19.9758 6.69381 17.9039 6.32194Z" />
-                  <path d="M17.0531 24.8625H16.8937C16.3625 24.8625 15.9375 25.2875 15.9375 25.8188C15.9375 26.35 16.3625 26.7751 16.8937 26.7751H17.0531C17.5844 26.7751 18.0094 26.35 18.0094 25.8188C18.0094 25.2875 17.5844 24.8625 17.0531 24.8625Z" />
-                </svg>
-              </div>
-              <div className="w-full">
-                <h3 className="text-dark mb-4 text-xl font-semibold  sm:text-2xl lg:text-xl xl:text-2xl">
-                  What is JANI AI
-                </h3>
-                <p className=" text-base lg:text-lg">
-                  JANI AI helps agribusinesses track their resources across the
-                  supply chain, generate reports easily, as well as record
-                  keeping management.
-                </p>
-              </div>
-            </div>
-            <div className="mb-12 flex lg:mb-[70px]">
-              <div className="mr-4 flex h-[50px] w-full max-w-[50px] items-center justify-center rounded-xl bg-primary text-white sm:mr-6 sm:h-[60px] sm:max-w-[60px]">
-                <svg
-                  width="32"
-                  height="32"
-                  viewBox="0 0 34 34"
-                  className="fill-current"
-                >
-                  <path d="M17.0008 0.690674C7.96953 0.690674 0.691406 7.9688 0.691406 17C0.691406 26.0313 7.96953 33.3625 17.0008 33.3625C26.032 33.3625 33.3633 26.0313 33.3633 17C33.3633 7.9688 26.032 0.690674 17.0008 0.690674ZM17.0008 31.5032C9.03203 31.5032 2.55078 24.9688 2.55078 17C2.55078 9.0313 9.03203 2.55005 17.0008 2.55005C24.9695 2.55005 31.5039 9.0313 31.5039 17C31.5039 24.9688 24.9695 31.5032 17.0008 31.5032Z" />
-                  <path d="M17.9039 6.32194C16.3633 6.05631 14.8227 6.48131 13.707 7.43756C12.5383 8.39381 11.8477 9.82819 11.8477 11.3688C11.8477 11.9532 11.9539 12.5376 12.1664 13.0688C12.3258 13.5469 12.857 13.8126 13.3352 13.6532C13.8133 13.4938 14.0789 12.9626 13.9195 12.4844C13.8133 12.1126 13.707 11.7938 13.707 11.3688C13.707 10.4126 14.132 9.50944 14.8758 8.87194C15.6195 8.23444 16.5758 7.96881 17.5852 8.18131C18.9133 8.39381 19.9758 9.50944 20.1883 10.7844C20.4539 12.3251 19.657 13.8126 18.2227 14.3969C16.8945 14.9282 16.0445 16.2563 16.0445 17.7969V21.1969C16.0445 21.7282 16.4695 22.1532 17.0008 22.1532C17.532 22.1532 17.957 21.7282 17.957 21.1969V17.7969C17.957 17.0532 18.382 16.3626 18.9664 16.1501C21.1977 15.2469 22.4727 12.9094 22.0477 10.4657C21.6758 8.39381 19.9758 6.69381 17.9039 6.32194Z" />
-                  <path d="M17.0531 24.8625H16.8937C16.3625 24.8625 15.9375 25.2875 15.9375 25.8188C15.9375 26.35 16.3625 26.7751 16.8937 26.7751H17.0531C17.5844 26.7751 18.0094 26.35 18.0094 25.8188C18.0094 25.2875 17.5844 24.8625 17.0531 24.8625Z" />
-                </svg>
-              </div>
-              <div className="w-full">
-                <h3 className="text-dark mb-4 text-xl font-semibold sm:text-2xl lg:text-xl xl:text-2xl">
-                  Why Open Source
-                </h3>
-                <p className=" text-base lg:text-lg">
-                  We believe that data and custom workflows are crucial to the
-                  the success of any business. JANI allows businesses to fork the
-                  repo so that they can build their custom workflow and features
-                  on top of our existing features.
-                </p>
-              </div>
-            </div>
-          </div>
-          <div className="w-full px-4 lg:w-1/2">
-            <div className="mb-12 flex lg:mb-[70px]">
-              <div className="mr-4 flex h-[50px] w-full max-w-[50px] items-center justify-center rounded-xl bg-primary text-white sm:mr-6 sm:h-[60px] sm:max-w-[60px]">
-                <svg
-                  width="32"
-                  height="32"
-                  viewBox="0 0 34 34"
-                  className="fill-current"
-                >
-                  <path d="M17.0008 0.690674C7.96953 0.690674 0.691406 7.9688 0.691406 17C0.691406 26.0313 7.96953 33.3625 17.0008 33.3625C26.032 33.3625 33.3633 26.0313 33.3633 17C33.3633 7.9688 26.032 0.690674 17.0008 0.690674ZM17.0008 31.5032C9.03203 31.5032 2.55078 24.9688 2.55078 17C2.55078 9.0313 9.03203 2.55005 17.0008 2.55005C24.9695 2.55005 31.5039 9.0313 31.5039 17C31.5039 24.9688 24.9695 31.5032 17.0008 31.5032Z" />
-                  <path d="M17.9039 6.32194C16.3633 6.05631 14.8227 6.48131 13.707 7.43756C12.5383 8.39381 11.8477 9.82819 11.8477 11.3688C11.8477 11.9532 11.9539 12.5376 12.1664 13.0688C12.3258 13.5469 12.857 13.8126 13.3352 13.6532C13.8133 13.4938 14.0789 12.9626 13.9195 12.4844C13.8133 12.1126 13.707 11.7938 13.707 11.3688C13.707 10.4126 14.132 9.50944 14.8758 8.87194C15.6195 8.23444 16.5758 7.96881 17.5852 8.18131C18.9133 8.39381 19.9758 9.50944 20.1883 10.7844C20.4539 12.3251 19.657 13.8126 18.2227 14.3969C16.8945 14.9282 16.0445 16.2563 16.0445 17.7969V21.1969C16.0445 21.7282 16.4695 22.1532 17.0008 22.1532C17.532 22.1532 17.957 21.7282 17.957 21.1969V17.7969C17.957 17.0532 18.382 16.3626 18.9664 16.1501C21.1977 15.2469 22.4727 12.9094 22.0477 10.4657C21.6758 8.39381 19.9758 6.69381 17.9039 6.32194Z" />
-                  <path d="M17.0531 24.8625H16.8937C16.3625 24.8625 15.9375 25.2875 15.9375 25.8188C15.9375 26.35 16.3625 26.7751 16.8937 26.7751H17.0531C17.5844 26.7751 18.0094 26.35 18.0094 25.8188C18.0094 25.2875 17.5844 24.8625 17.0531 24.8625Z" />
-                </svg>
-              </div>
-              <div className="w-full">
-                <h3 className="text-dark mb-4 text-xl font-semibold sm:text-2xl lg:text-xl xl:text-2xl">
-                  Why you should care
-                </h3>
-                <p className=" text-base lg:text-lg">
-                  We are what we eat. Giving transparency to your clients on
-                  what processes and input go into the food they buy will increase loyalty, trust, and your brand
-                  value.
-                </p>
-              </div>
-            </div>
-            <div className="mb-12 flex lg:mb-[70px]">
-              <div className="mr-4 flex h-[50px] w-full max-w-[50px] items-center justify-center rounded-xl bg-primary text-white sm:mr-6 sm:h-[60px] sm:max-w-[60px]">
-                <svg
-                  width="32"
-                  height="32"
-                  viewBox="0 0 34 34"
-                  className="fill-current"
-                >
-                  <path d="M17.0008 0.690674C7.96953 0.690674 0.691406 7.9688 0.691406 17C0.691406 26.0313 7.96953 33.3625 17.0008 33.3625C26.032 33.3625 33.3633 26.0313 33.3633 17C33.3633 7.9688 26.032 0.690674 17.0008 0.690674ZM17.0008 31.5032C9.03203 31.5032 2.55078 24.9688 2.55078 17C2.55078 9.0313 9.03203 2.55005 17.0008 2.55005C24.9695 2.55005 31.5039 9.0313 31.5039 17C31.5039 24.9688 24.9695 31.5032 17.0008 31.5032Z" />
-                  <path d="M17.9039 6.32194C16.3633 6.05631 14.8227 6.48131 13.707 7.43756C12.5383 8.39381 11.8477 9.82819 11.8477 11.3688C11.8477 11.9532 11.9539 12.5376 12.1664 13.0688C12.3258 13.5469 12.857 13.8126 13.3352 13.6532C13.8133 13.4938 14.0789 12.9626 13.9195 12.4844C13.8133 12.1126 13.707 11.7938 13.707 11.3688C13.707 10.4126 14.132 9.50944 14.8758 8.87194C15.6195 8.23444 16.5758 7.96881 17.5852 8.18131C18.9133 8.39381 19.9758 9.50944 20.1883 10.7844C20.4539 12.3251 19.657 13.8126 18.2227 14.3969C16.8945 14.9282 16.0445 16.2563 16.0445 17.7969V21.1969C16.0445 21.7282 16.4695 22.1532 17.0008 22.1532C17.532 22.1532 17.957 21.7282 17.957 21.1969V17.7969C17.957 17.0532 18.382 16.3626 18.9664 16.1501C21.1977 15.2469 22.4727 12.9094 22.0477 10.4657C21.6758 8.39381 19.9758 6.69381 17.9039 6.32194Z" />
-                  <path d="M17.0531 24.8625H16.8937C16.3625 24.8625 15.9375 25.2875 15.9375 25.8188C15.9375 26.35 16.3625 26.7751 16.8937 26.7751H17.0531C17.5844 26.7751 18.0094 26.35 18.0094 25.8188C18.0094 25.2875 17.5844 24.8625 17.0531 24.8625Z" />
-                </svg>
-              </div>
-              <div className="w-full">
-                <h3 className="text-dark mb-4 text-xl font-semibold sm:text-2xl lg:text-xl xl:text-2xl">
-                  How can I self-host?
-                </h3>
-                <p className=" text-base lg:text-lg">
-                  Each software created comes with a dedicated manual on
-                  how you can self-host. JANI AI guarantees 99% uptime for all
-                  the software shipped so that you can keep on focusing on your
-                  business operations.
-                </p>
-              </div>
-            </div>
-          </div>
-        </div>
-      </div>
-      <div>
-        <span className="absolute left-4 top-4 -z-[1]">
-          <svg
-            width="48"
-            height="134"
-            viewBox="0 0 48 134"
-            fill="none"
-            xmlns="http://www.w3.org/2000/svg"
-          >
-            <circle
-              cx="45.6673"
-              cy="132"
-              r="1.66667"
-              transform="rotate(180 45.6673 132)"
-              fill="#13C296"
-            />
-            <circle
-              cx="45.6673"
-              cy="117.333"
-              r="1.66667"
-              transform="rotate(180 45.6673 117.333)"
-              fill="#13C296"
-            />
-            <circle
-              cx="45.6673"
-              cy="102.667"
-              r="1.66667"
-              transform="rotate(180 45.6673 102.667)"
-              fill="#13C296"
-            />
-            <circle
-              cx="45.6673"
-              cy="88.0001"
-              r="1.66667"
-              transform="rotate(180 45.6673 88.0001)"
-              fill="#13C296"
-            />
-            <circle
-              cx="45.6673"
-              cy="73.3335"
-              r="1.66667"
-              transform="rotate(180 45.6673 73.3335)"
-              fill="#13C296"
-            />
-            <circle
-              cx="45.6673"
-              cy="45.0001"
-              r="1.66667"
-              transform="rotate(180 45.6673 45.0001)"
-              fill="#13C296"
-            />
-            <circle
-              cx="45.6673"
-              cy="16.0001"
-              r="1.66667"
-              transform="rotate(180 45.6673 16.0001)"
-              fill="#13C296"
-            />
-            <circle
-              cx="45.6673"
-              cy="59.0001"
-              r="1.66667"
-              transform="rotate(180 45.6673 59.0001)"
-              fill="#13C296"
-            />
-            <circle
-              cx="45.6673"
-              cy="30.6668"
-              r="1.66667"
-              transform="rotate(180 45.6673 30.6668)"
-              fill="#13C296"
-            />
-            <circle
-              cx="45.6673"
-              cy="1.66683"
-              r="1.66667"
-              transform="rotate(180 45.6673 1.66683)"
-              fill="#13C296"
-            />
-            <circle
-              cx="31.0013"
-              cy="132"
-              r="1.66667"
-              transform="rotate(180 31.0013 132)"
-              fill="#13C296"
-            />
-            <circle
-              cx="31.0013"
-              cy="117.333"
-              r="1.66667"
-              transform="rotate(180 31.0013 117.333)"
-              fill="#13C296"
-            />
-            <circle
-              cx="31.0013"
-              cy="102.667"
-              r="1.66667"
-              transform="rotate(180 31.0013 102.667)"
-              fill="#13C296"
-            />
-            <circle
-              cx="31.0013"
-              cy="88.0001"
-              r="1.66667"
-              transform="rotate(180 31.0013 88.0001)"
-              fill="#13C296"
-            />
-            <circle
-              cx="31.0013"
-              cy="73.3335"
-              r="1.66667"
-              transform="rotate(180 31.0013 73.3335)"
-              fill="#13C296"
-            />
-            <circle
-              cx="31.0013"
-              cy="45.0001"
-              r="1.66667"
-              transform="rotate(180 31.0013 45.0001)"
-              fill="#13C296"
-            />
-            <circle
-              cx="31.0013"
-              cy="16.0001"
-              r="1.66667"
-              transform="rotate(180 31.0013 16.0001)"
-              fill="#13C296"
-            />
-            <circle
-              cx="31.0013"
-              cy="59.0001"
-              r="1.66667"
-              transform="rotate(180 31.0013 59.0001)"
-              fill="#13C296"
-            />
-            <circle
-              cx="31.0013"
-              cy="30.6668"
-              r="1.66667"
-              transform="rotate(180 31.0013 30.6668)"
-              fill="#13C296"
-            />
-            <circle
-              cx="31.0013"
-              cy="1.66683"
-              r="1.66667"
-              transform="rotate(180 31.0013 1.66683)"
-              fill="#13C296"
-            />
-            <circle
-              cx="16.3333"
-              cy="132"
-              r="1.66667"
-              transform="rotate(180 16.3333 132)"
-              fill="#13C296"
-            />
-            <circle
-              cx="16.3333"
-              cy="117.333"
-              r="1.66667"
-              transform="rotate(180 16.3333 117.333)"
-              fill="#13C296"
-            />
-            <circle
-              cx="16.3333"
-              cy="102.667"
-              r="1.66667"
-              transform="rotate(180 16.3333 102.667)"
-              fill="#13C296"
-            />
-            <circle
-              cx="16.3333"
-              cy="88.0001"
-              r="1.66667"
-              transform="rotate(180 16.3333 88.0001)"
-              fill="#13C296"
-            />
-            <circle
-              cx="16.3333"
-              cy="73.3335"
-              r="1.66667"
-              transform="rotate(180 16.3333 73.3335)"
-              fill="#13C296"
-            />
-            <circle
-              cx="16.3333"
-              cy="45.0001"
-              r="1.66667"
-              transform="rotate(180 16.3333 45.0001)"
-              fill="#13C296"
-            />
-            <circle
-              cx="16.3333"
-              cy="16.0001"
-              r="1.66667"
-              transform="rotate(180 16.3333 16.0001)"
-              fill="#13C296"
-            />
-            <circle
-              cx="16.3333"
-              cy="59.0001"
-              r="1.66667"
-              transform="rotate(180 16.3333 59.0001)"
-              fill="#13C296"
-            />
-            <circle
-              cx="16.3333"
-              cy="30.6668"
-              r="1.66667"
-              transform="rotate(180 16.3333 30.6668)"
-              fill="#13C296"
-            />
-            <circle
-              cx="16.3333"
-              cy="1.66683"
-              r="1.66667"
-              transform="rotate(180 16.3333 1.66683)"
-              fill="#13C296"
-            />
-            <circle
-              cx="1.66732"
-              cy="132"
-              r="1.66667"
-              transform="rotate(180 1.66732 132)"
-              fill="#13C296"
-            />
-            <circle
-              cx="1.66732"
-              cy="117.333"
-              r="1.66667"
-              transform="rotate(180 1.66732 117.333)"
-              fill="#13C296"
-            />
-            <circle
-              cx="1.66732"
-              cy="102.667"
-              r="1.66667"
-              transform="rotate(180 1.66732 102.667)"
-              fill="#13C296"
-            />
-            <circle
-              cx="1.66732"
-              cy="88.0001"
-              r="1.66667"
-              transform="rotate(180 1.66732 88.0001)"
-              fill="#13C296"
-            />
-            <circle
-              cx="1.66732"
-              cy="73.3335"
-              r="1.66667"
-              transform="rotate(180 1.66732 73.3335)"
-              fill="#13C296"
-            />
-            <circle
-              cx="1.66732"
-              cy="45.0001"
-              r="1.66667"
-              transform="rotate(180 1.66732 45.0001)"
-              fill="#13C296"
-            />
-            <circle
-              cx="1.66732"
-              cy="16.0001"
-              r="1.66667"
-              transform="rotate(180 1.66732 16.0001)"
-              fill="#13C296"
-            />
-            <circle
-              cx="1.66732"
-              cy="59.0001"
-              r="1.66667"
-              transform="rotate(180 1.66732 59.0001)"
-              fill="#13C296"
-            />
-            <circle
-              cx="1.66732"
-              cy="30.6668"
-              r="1.66667"
-              transform="rotate(180 1.66732 30.6668)"
-              fill="#13C296"
-            />
-            <circle
-              cx="1.66732"
-              cy="1.66683"
-              r="1.66667"
-              transform="rotate(180 1.66732 1.66683)"
-              fill="#13C296"
-            />
-          </svg>
-        </span>
-        <span className="absolute bottom-4 right-4 -z-[1]">
-          <svg
-            width="48"
-            height="134"
-            viewBox="0 0 48 134"
-            fill="none"
-            xmlns="http://www.w3.org/2000/svg"
-          >
-            <circle
-              cx="45.6673"
-              cy="132"
-              r="1.66667"
-              transform="rotate(180 45.6673 132)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="45.6673"
-              cy="117.333"
-              r="1.66667"
-              transform="rotate(180 45.6673 117.333)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="45.6673"
-              cy="102.667"
-              r="1.66667"
-              transform="rotate(180 45.6673 102.667)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="45.6673"
-              cy="88.0001"
-              r="1.66667"
-              transform="rotate(180 45.6673 88.0001)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="45.6673"
-              cy="73.3333"
-              r="1.66667"
-              transform="rotate(180 45.6673 73.3333)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="45.6673"
-              cy="45.0001"
-              r="1.66667"
-              transform="rotate(180 45.6673 45.0001)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="45.6673"
-              cy="16.0001"
-              r="1.66667"
-              transform="rotate(180 45.6673 16.0001)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="45.6673"
-              cy="59.0001"
-              r="1.66667"
-              transform="rotate(180 45.6673 59.0001)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="45.6673"
-              cy="30.6668"
-              r="1.66667"
-              transform="rotate(180 45.6673 30.6668)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="45.6673"
-              cy="1.66683"
-              r="1.66667"
-              transform="rotate(180 45.6673 1.66683)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="31.0006"
-              cy="132"
-              r="1.66667"
-              transform="rotate(180 31.0006 132)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="31.0006"
-              cy="117.333"
-              r="1.66667"
-              transform="rotate(180 31.0006 117.333)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="31.0006"
-              cy="102.667"
-              r="1.66667"
-              transform="rotate(180 31.0006 102.667)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="31.0006"
-              cy="88.0001"
-              r="1.66667"
-              transform="rotate(180 31.0006 88.0001)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="31.0008"
-              cy="73.3333"
-              r="1.66667"
-              transform="rotate(180 31.0008 73.3333)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="31.0008"
-              cy="45.0001"
-              r="1.66667"
-              transform="rotate(180 31.0008 45.0001)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="31.0008"
-              cy="16.0001"
-              r="1.66667"
-              transform="rotate(180 31.0008 16.0001)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="31.0008"
-              cy="59.0001"
-              r="1.66667"
-              transform="rotate(180 31.0008 59.0001)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="31.0008"
-              cy="30.6668"
-              r="1.66667"
-              transform="rotate(180 31.0008 30.6668)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="31.0008"
-              cy="1.66683"
-              r="1.66667"
-              transform="rotate(180 31.0008 1.66683)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="16.3341"
-              cy="132"
-              r="1.66667"
-              transform="rotate(180 16.3341 132)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="16.3341"
-              cy="117.333"
-              r="1.66667"
-              transform="rotate(180 16.3341 117.333)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="16.3341"
-              cy="102.667"
-              r="1.66667"
-              transform="rotate(180 16.3341 102.667)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="16.3341"
-              cy="88.0001"
-              r="1.66667"
-              transform="rotate(180 16.3341 88.0001)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="16.3338"
-              cy="73.3333"
-              r="1.66667"
-              transform="rotate(180 16.3338 73.3333)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="16.3338"
-              cy="45.0001"
-              r="1.66667"
-              transform="rotate(180 16.3338 45.0001)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="16.3338"
-              cy="16.0001"
-              r="1.66667"
-              transform="rotate(180 16.3338 16.0001)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="16.3338"
-              cy="59.0001"
-              r="1.66667"
-              transform="rotate(180 16.3338 59.0001)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="16.3338"
-              cy="30.6668"
-              r="1.66667"
-              transform="rotate(180 16.3338 30.6668)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="16.3338"
-              cy="1.66683"
-              r="1.66667"
-              transform="rotate(180 16.3338 1.66683)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="1.66732"
-              cy="132"
-              r="1.66667"
-              transform="rotate(180 1.66732 132)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="1.66732"
-              cy="117.333"
-              r="1.66667"
-              transform="rotate(180 1.66732 117.333)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="1.66732"
-              cy="102.667"
-              r="1.66667"
-              transform="rotate(180 1.66732 102.667)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="1.66732"
-              cy="88.0001"
-              r="1.66667"
-              transform="rotate(180 1.66732 88.0001)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="1.66732"
-              cy="73.3333"
-              r="1.66667"
-              transform="rotate(180 1.66732 73.3333)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="1.66732"
-              cy="45.0001"
-              r="1.66667"
-              transform="rotate(180 1.66732 45.0001)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="1.66732"
-              cy="16.0001"
-              r="1.66667"
-              transform="rotate(180 1.66732 16.0001)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="1.66732"
-              cy="59.0001"
-              r="1.66667"
-              transform="rotate(180 1.66732 59.0001)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="1.66732"
-              cy="30.6668"
-              r="1.66667"
-              transform="rotate(180 1.66732 30.6668)"
-              fill="#3758F9"
-            />
-            <circle
-              cx="1.66732"
-              cy="1.66683"
-              r="1.66667"
-              transform="rotate(180 1.66732 1.66683)"
-              fill="#3758F9"
-            />
-          </svg>
-        </span>
-      </div>
-    </section>
-  );
-}
