--- apps/nextjs/src/components/landingPage/oldComponents/content.tsx
@@ -1,82 +1,0 @@
-
-export const FEATURES_LIST = [
-  {
-    title: "Farmer Management",
-    description: "Manage records for your farmers",
-    // icon: ChartLine,
-    href: "#",
-  },
-  {
-    title: "Traceability Reports",
-    description: "Reports that map out your supply chain",
-    // icon: Hyperlink,
-    href: "#",
-  },
-  {
-    title: "QR Codes",
-    description: "Gorgeous QR codes for your traceability reports",
-    // icon: QRCode,
-    href: "#",
-  },
-  {
-    title: "Personalization",
-    description: "Personalize Your Reports",
-    // icon: Paintbrush,
-    href: "#",
-  },
-  {
-    title: "Collaboration",
-    description: "Collaborate With Your Team",
-    // icon: Users,
-    href: "#",
-  },
-  
-];
-
-export const CUSTOMER_STORIES = [
-  {
-    // icon: Raycast,
-    iconClassName: "group-hover:text-[#FF6363]",
-    title: "Nestle",
-    description:
-      "How Nestle tracks their resources from remote farmers in Ivory Coast",
-    href: "/customers/raycast",
-  },
-  
-];
-
-
-
-
-export const RESOURCES = [
-  {
-    // icon: Headset,
-    title: "Help Center",
-    href: "/help",
-  },
-  {
-    // icon: Book2,
-    title: "Docs",
-    href: "/docs",
-  },
-  {
-    // icon: Blog,
-    title: "Blog",
-    href: "/blog",
-  },
-  {
-    // icon: BulletList,
-    title: "Changelog",
-    href: "/changelog",
-  },
-  {
-    // icon: Stars2,
-    title: "Customers",
-    href: "/customers",
-  },
-  {
-    // icon: HexadecagonStar,
-    title: "Brand",
-    href: "/brand",
-  },
-];
