--- apps/nextjs/src/components/landingPage/oldComponents/index.ts
@@ -1,6 +1,0 @@
-export { default as Features } from "./Features";
-export { default as HeroSection } from "./HeroSection";
-export { default as About } from "./About";
-export { default as Call2Action } from "./Call2Action";
-export { default as Pricing } from "./Pricing";
-export {default as Faq} from "./Faq"
